HSEARCH-4535 Move PredicateDefinition/PredicateDefinitionContext to a more sensible package
