+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/document/model/dsl/impl/AbstractLuceneIndexCompositeNodeBuilder.java
@@ -27,7 +27,7 @@ import org.hibernate.search.engine.backend.document.model.dsl.spi.IndexComposite
 import org.hibernate.search.engine.backend.document.model.spi.IndexFieldInclusion;
 import org.hibernate.search.engine.backend.types.IndexFieldType;
 import org.hibernate.search.engine.backend.types.ObjectStructure;
-import org.hibernate.search.engine.search.predicate.factories.PredicateDefinition;
+import org.hibernate.search.engine.search.predicate.definition.PredicateDefinition;
 import org.hibernate.search.engine.search.predicate.spi.PredicateTypeKeys;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
