+++ engine/src/main/java/org/hibernate/search/engine/backend/document/model/dsl/spi/IndexCompositeNodeBuilder.java
@@ -13,7 +13,7 @@ import org.hibernate.search.engine.backend.types.ObjectStructure;
 import org.hibernate.search.engine.backend.document.model.spi.IndexFieldInclusion;
 import org.hibernate.search.engine.backend.types.IndexFieldType;
 import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaNamedPredicateOptionsStep;
-import org.hibernate.search.engine.search.predicate.factories.PredicateDefinition;
+import org.hibernate.search.engine.search.predicate.definition.PredicateDefinition;
 
 public interface IndexCompositeNodeBuilder extends IndexSchemaBuildContext {
